Move AnsibleJSONEncoder to module_utils, consolidate instances
